Restructure directories root
