+++ mediark/application/domain/services/__init__.py
@@ -2,3 +2,4 @@ from .auth import *
 from .file_store import *
 from .id_service import *
 from .cache_service import *
+from .repositories import *
