+++ mediark/application/domain/services/repositories/__init__.py
@@ -1,3 +1,3 @@
-from ..common import QueryDomain
+from ...common import QueryDomain
 from .memory_model_repositories import (
     MediaRepository, MemoryMediaRepository)
