+++ mediark/application/operation/managers/__init__.py
