+++ mediark/presentation/rest/middleware/errors.py
@@ -4,7 +4,7 @@ from traceback import format_tb
 from typing import Callable, Dict, Any
 from aiohttp import web
 from injectark import Injectark
-from ....application.managers import SessionManager
+from ....application.operation.managers import SessionManager
 from ....integration.core.suppliers.common.tenancy import TenantSupplier
